mybot/slack/requirements.txt
slackclient==1.0.2
PyYAML==3.11
